+++ apps/app/src/components/TreeItem/NewPageInput/index.ts
@@ -1,0 +1,1 @@
+export * from './use-new-page-input';
