+++ apps/app/src/components/TreeItem/index.ts
@@ -1,6 +1,6 @@
 export * from './interfaces';
 
-export * from './UseNewPageInput';
+export * from './NewPageInput';
 export * from './ItemNode';
 export * from './SimpleItem';
 export * from './NotDraggableForClosableTextInput';
